Consolidate variant code
