+++ src/compiler/typescript_to_rust/azle_generate/src/ts_ast/azle_type/azle_type_lit/azle_type_element/mod.rs
@@ -1,8 +1,6 @@
 use swc_common::SourceMap;
 use swc_ecma_ast::TsTypeElement;
 
-use cdk_framework::act::node::candid::{record, variant};
-
 use self::{
     azle_method_signature::AzleMethodSignature, azle_property_signature::AzlePropertySignature,
 };
@@ -46,15 +44,7 @@ impl AzleTypeElement<'_> {
         }
     }
 
-    pub(super) fn to_variant_member(&self) -> variant::Member {
-        let ts_property_signature = match self.as_azle_property_signature() {
-            Some(ts_property_signature) => ts_property_signature,
-            None => panic!("{}", self.variant_property_signature_error()),
-        };
-        ts_property_signature.to_variant_member()
-    }
-
-    pub(self) fn get_source_map(&self) -> &SourceMap {
+    pub fn get_source_map(&self) -> &SourceMap {
         match self {
             AzleTypeElement::AzlePropertySignature(azle_property_signature) => {
                 azle_property_signature.source_map
