+++ src/compiler/typescript_to_rust/azle_generate/src/ts_ast/azle_type/azle_type_lit/mod.rs
@@ -15,23 +15,3 @@ pub struct AzleTypeLit<'a> {
     pub ts_type_lit: TsTypeLit,
     pub source_map: &'a SourceMap,
 }
-
-impl AzleTypeLit<'_> {
-    pub fn to_variant(&self) -> Variant {
-        let members: Vec<variant::Member> = self
-            .ts_type_lit
-            .members
-            .iter()
-            .map(|member| {
-                let azle_member =
-                    AzleTypeElement::from_ts_type_element(member.clone(), self.source_map);
-                azle_member.to_variant_member()
-            })
-            .collect();
-
-        Variant {
-            name: None,
-            members,
-        }
-    }
-}
